semi final
added option to delete node, information frame to view all the edge's information, updated arrow heads(they are not lousy now), made look and feel better, made option to use the algorithms, but they are not implemented yet, changed code to look nicer and added comments. In my opinion the gui is done.